Major Update: Fusion360 redesign
Redesigned the joystick from scratch in Fusion 360. Modelling in 3d polygonal app (Modo) was OK, but as the complexity grew, it became unmanageably large. Simple operations such as moving a screw-hole was taking 10+ minutes just to process.

This redesign is more game-accurate and works better. I redesigned the PCB's for this new design but sadly, they just don't quite work - the light for the HAT switches simply doesn't shine through, so I will be redoing the two PCBs at the top of teh joystick again with side-mounted LEDs to better light.

Separately, the two info panels (the palm panel and the switches panel) have been redesigned in Fusion360 as well, and are nearing completion. Once finished I'll add the prints to this directory as well.